Threading WIP -> Can't get it to make the requests. Maybe there is something that blocks it?
Deadlock error was due to not passing an argument to ::join

More threading WIP, this is close, but there is something to do with the pool loop that it blocks the remaining requests. Maybe it tries to divvy requests as much as possible and then relies on join to make them live or not.

use ruby 222

more threading wip

Threading is still not working. I can write the file, but there are a couple of rows missing.

Add testing file